pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.13.4",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "rich>=14.3.3",
    "streamlit>=1.56.0",
]
